pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rio-protocol"
version = "1.0.0"
description = "RIO — Runtime Intelligence Orchestration: Fail-closed execution governance with cryptographic audit"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "Brian K. Rasmussen"},
]
keywords = ["governance", "audit", "compliance", "ai-safety", "execution-control"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "cryptography>=41.0.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0",
    "jinja2>=3.1.0",
    "plotly>=5.18.0",
    "requests>=2.31.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
]

[project.scripts]
rio-init = "scripts.init_rio:main"
rio-run = "scripts.run_all:main"
rio-admin = "scripts.create_admin_user:main"

[project.urls]
Homepage = "https://github.com/bkr1297-RIO/rio-protocol"
Repository = "https://github.com/bkr1297-RIO/rio-protocol"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.setuptools.packages.find]
exclude = ["tests*", "whitepaper*", "examples*"]
